requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
google-genai


langchain
langchain_community
langchain-google-genai
python-dotenv
streamlit
langchain_experimental
sentence-transformers
langchain_chroma
langchainhub
pypdf
rapidocr-onnxruntime